feat(store): handle errors in shareData slice
